Threaded Download and Watch
Download and watch works by downloading the stream in a thread to a local
directory and playing the local file simultaneously. It currently uses
environment variables to communicate with the thread as a new python
interpreter seems to be spawned on every plugin command so the threads
are not accessible for subsequent commands.

Anotther option to communicate with the download thread is to use temp
files. This might have to be added to support windows.

Community testing of this patch has revealed that on Windows, this
patch only works with latest nightly builds of XBMC and it does not
work with the stable releases.

TODO: add checks for OS and xbmc version and enable/disable download and
watch function accordingly.
TODO: Add option to communicate with download thread by temp files (for
Windows users)
TODO: Better support for multiple parts (download second as soon as
first part is complete)